Extract components/orms/ out of Offer page
Setting them up for reuse on Ask page.

Also:
* Moves AuthTokenInput and DeleteButton under forms/
* Moves AboutSection under pages/orientation/ since only used there.
* Fills out components/index.js to ease use of importing.
