+++ app/javascript/components/forms/AuthTokenInput.vue
