+++ app/javascript/components/forms/CategoryFields.vue
