+++ app/javascript/components/forms/index.js
@@ -1,0 +1,18 @@
+import AuthTokenInput   from './AuthTokenInput'
+import CategoryFields   from './CategoryFields'
+import ContactFields    from './ContactFields'
+import DeleteButton     from './DeleteButton'
+import LocationFields   from './LocationFields'
+import ServiceAreaField from './ServiceAreaField'
+import SpacerField      from './SpacerField'
+
+export {
+  AuthTokenInput,
+  CategoryFields,
+  ContactFields,
+  DeleteButton,
+  LocationFields,
+  ServiceAreaField,
+  SpacerField,
+}
+
